+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### New features
 
+- Added `\Wizaplace\SDK\Catalog\Product::getAvailableSince`
 - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductFilter`
 - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductList`
 - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::getListMultiVendorProduct`
+
+## 1.49.0
 
 ## 1.48.0